Add first signing tests
