+++ client/tests/input_selection/native_tokens.rs
@@ -10,7 +10,7 @@ use iota_client::{
 };
 use primitive_types::U256;
 
-use crate::input_selection::{
+use crate::{
     addresses, build_inputs, build_outputs, is_remainder_or_return, unsorted_eq, Build::Basic,
     BECH32_ADDRESS_ED25519_0, TOKEN_ID_1, TOKEN_ID_2,
 };
@@ -28,11 +28,13 @@ fn two_native_tokens_one_needed() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100), (TOKEN_ID_2, 100)]),
+            None,
             None,
             None,
             None,
@@ -43,6 +45,7 @@ fn two_native_tokens_one_needed() {
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 150)]),
+        None,
         None,
         None,
         None,
@@ -86,11 +89,13 @@ fn two_native_tokens_both_needed_plus_remainder() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100), (TOKEN_ID_2, 100)]),
+            None,
             None,
             None,
             None,
@@ -101,6 +106,7 @@ fn two_native_tokens_both_needed_plus_remainder() {
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 150), (TOKEN_ID_2, 100)]),
+        None,
         None,
         None,
         None,
@@ -144,20 +150,23 @@ fn three_inputs_two_needed_plus_remainder() {
             None,
             None,
             None,
-        ),
-        Basic(
-            1_000_000,
-            BECH32_ADDRESS_ED25519_0,
-            Some(vec![(TOKEN_ID_1, 100)]),
-            None,
-            None,
-            None,
             None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100)]),
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(
+            1_000_000,
+            BECH32_ADDRESS_ED25519_0,
+            Some(vec![(TOKEN_ID_1, 100)]),
+            None,
             None,
             None,
             None,
@@ -168,6 +177,7 @@ fn three_inputs_two_needed_plus_remainder() {
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 120)]),
+        None,
         None,
         None,
         None,
@@ -211,20 +221,23 @@ fn three_inputs_two_needed_no_remainder() {
             None,
             None,
             None,
-        ),
-        Basic(
-            1_000_000,
-            BECH32_ADDRESS_ED25519_0,
-            Some(vec![(TOKEN_ID_1, 100)]),
-            None,
-            None,
-            None,
             None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100)]),
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(
+            1_000_000,
+            BECH32_ADDRESS_ED25519_0,
+            Some(vec![(TOKEN_ID_1, 100)]),
+            None,
             None,
             None,
             None,
@@ -235,6 +248,7 @@ fn three_inputs_two_needed_no_remainder() {
         2_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 200)]),
+        None,
         None,
         None,
         None,
@@ -266,11 +280,13 @@ fn insufficient_native_tokens_one_input() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 150)]),
+        None,
         None,
         None,
         None,
@@ -307,20 +323,23 @@ fn insufficient_native_tokens_three_inputs() {
             None,
             None,
             None,
-        ),
-        Basic(
-            1_000_000,
-            BECH32_ADDRESS_ED25519_0,
-            Some(vec![(TOKEN_ID_1, 100)]),
-            None,
-            None,
-            None,
             None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100)]),
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(
+            1_000_000,
+            BECH32_ADDRESS_ED25519_0,
+            Some(vec![(TOKEN_ID_1, 100)]),
+            None,
             None,
             None,
             None,
@@ -331,6 +350,7 @@ fn insufficient_native_tokens_three_inputs() {
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 301)]),
+        None,
         None,
         None,
         None,
@@ -366,11 +386,13 @@ fn burn_and_send_at_the_same_time() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 50), (TOKEN_ID_2, 100)]),
+        None,
         None,
         None,
         None,
@@ -414,6 +436,7 @@ fn burn_one_input_no_output() {
         None,
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -449,11 +472,13 @@ fn burn_two_inputs_no_output() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100), (TOKEN_ID_2, 100)]),
+            None,
             None,
             None,
             None,
@@ -494,6 +519,7 @@ fn burn_one_input_two_tokens_no_output() {
         None,
         None,
         None,
+        None,
     )]);
 
     let selected = InputSelection::new(
@@ -529,11 +555,13 @@ fn multiple_native_tokens_1() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100), (TOKEN_ID_2, 100)]),
+            None,
             None,
             None,
             None,
@@ -544,6 +572,7 @@ fn multiple_native_tokens_1() {
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 100)]),
+        None,
         None,
         None,
         None,
@@ -577,11 +606,13 @@ fn multiple_native_tokens_2() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100), (TOKEN_ID_2, 100)]),
+            None,
             None,
             None,
             None,
@@ -592,6 +623,7 @@ fn multiple_native_tokens_2() {
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 150)]),
+        None,
         None,
         None,
         None,
@@ -631,11 +663,13 @@ fn multiple_native_tokens_3() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 100), (TOKEN_ID_2, 100)]),
+            None,
             None,
             None,
             None,
@@ -646,6 +680,7 @@ fn multiple_native_tokens_3() {
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 200)]),
+        None,
         None,
         None,
         None,
@@ -684,11 +719,13 @@ fn insufficient_native_tokens() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 150)]),
+        None,
         None,
         None,
         None,
@@ -724,11 +761,13 @@ fn insufficient_native_tokens_2() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 150)]),
+        None,
         None,
         None,
         None,
@@ -764,11 +803,13 @@ fn insufficient_amount_for_remainder() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 50)]),
+        None,
         None,
         None,
         None,
@@ -804,11 +845,13 @@ fn single_output_native_token_no_remainder() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 100)]),
+        None,
         None,
         None,
         None,
@@ -840,11 +883,13 @@ fn single_output_native_token_remainder_1() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 50)]),
+        None,
         None,
         None,
         None,
@@ -883,11 +928,13 @@ fn single_output_native_token_remainder_2() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 100)]),
+        None,
         None,
         None,
         None,
@@ -927,6 +974,7 @@ fn two_basic_outputs_1() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
@@ -936,11 +984,13 @@ fn two_basic_outputs_1() {
             None,
             None,
             None,
+            None,
         ),
     ]);
     let outputs = build_outputs(vec![Basic(
         500_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -982,11 +1032,13 @@ fn two_basic_outputs_2() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 200)]),
+            None,
             None,
             None,
             None,
@@ -997,6 +1049,7 @@ fn two_basic_outputs_2() {
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 50)]),
+        None,
         None,
         None,
         None,
@@ -1037,11 +1090,13 @@ fn two_basic_outputs_3() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 200)]),
+            None,
             None,
             None,
             None,
@@ -1052,6 +1107,7 @@ fn two_basic_outputs_3() {
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 75)]),
+        None,
         None,
         None,
         None,
@@ -1092,11 +1148,13 @@ fn two_basic_outputs_4() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 200)]),
+            None,
             None,
             None,
             None,
@@ -1107,6 +1165,7 @@ fn two_basic_outputs_4() {
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 100)]),
+        None,
         None,
         None,
         None,
@@ -1147,11 +1206,13 @@ fn two_basic_outputs_5() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 200)]),
+            None,
             None,
             None,
             None,
@@ -1162,6 +1223,7 @@ fn two_basic_outputs_5() {
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 100)]),
+        None,
         None,
         None,
         None,
@@ -1202,11 +1264,13 @@ fn two_basic_outputs_6() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 200)]),
+            None,
             None,
             None,
             None,
@@ -1217,6 +1281,7 @@ fn two_basic_outputs_6() {
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 250)]),
+        None,
         None,
         None,
         None,
@@ -1256,11 +1321,13 @@ fn two_basic_outputs_7() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 200)]),
+            None,
             None,
             None,
             None,
@@ -1271,6 +1338,7 @@ fn two_basic_outputs_7() {
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 300)]),
+        None,
         None,
         None,
         None,
@@ -1310,11 +1378,13 @@ fn two_basic_outputs_8() {
             None,
             None,
             None,
+            None,
         ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             Some(vec![(TOKEN_ID_1, 200)]),
+            None,
             None,
             None,
             None,
@@ -1325,6 +1395,7 @@ fn two_basic_outputs_8() {
         500_000,
         BECH32_ADDRESS_ED25519_0,
         Some(vec![(TOKEN_ID_1, 350)]),
+        None,
         None,
         None,
         None,
