+++ client/tests/input_selection/nft_outputs.rs
@@ -13,7 +13,7 @@ use iota_client::{
     Error,
 };
 
-use crate::input_selection::{
+use crate::{
     addresses, build_inputs, build_outputs, is_remainder_or_return, unsorted_eq,
     Build::{Basic, Nft},
     BECH32_ADDRESS_ALIAS_1, BECH32_ADDRESS_ED25519_0, BECH32_ADDRESS_ED25519_1, BECH32_ADDRESS_NFT_1, NFT_ID_0,
@@ -33,11 +33,13 @@ fn input_nft_eq_output_nft() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -70,12 +72,14 @@ fn transition_nft_id_zero() {
         None,
         None,
         None,
+        None,
     )]);
     let nft_id = NftId::from(inputs[0].output_id());
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -108,10 +112,12 @@ fn input_amount_lt_output_amount() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         2_000_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -150,10 +156,12 @@ fn basic_output_with_nft_input() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         2_000_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -188,11 +196,13 @@ fn mint_nft() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_0,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -234,10 +244,12 @@ fn burn_nft() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         2_000_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -272,11 +284,13 @@ fn not_enough_storage_deposit_for_remainder() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -313,11 +327,13 @@ fn missing_input_for_nft_output() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_2,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -351,11 +367,13 @@ fn missing_input_for_nft_output_but_created() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_0,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -379,16 +397,35 @@ fn nft_in_output_and_sender() {
     let nft_id_1 = NftId::from_str(NFT_ID_1).unwrap();
 
     let inputs = build_inputs(vec![
-        Nft(1_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
-        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None),
+        Nft(
+            1_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
     let outputs = build_outputs(vec![
-        Nft(1_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
+        Nft(
+            1_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
         Basic(
             1_000_000,
             BECH32_ADDRESS_ED25519_0,
             None,
             Some(BECH32_ADDRESS_NFT_1),
+            None,
             None,
             None,
             None,
@@ -429,6 +466,7 @@ fn missing_ed25519_sender() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -436,6 +474,7 @@ fn missing_ed25519_sender() {
         BECH32_ADDRESS_ED25519_0,
         None,
         Some(BECH32_ADDRESS_ED25519_1),
+        None,
         None,
         None,
     )]);
@@ -467,6 +506,7 @@ fn missing_ed25519_issuer_created() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -475,6 +515,7 @@ fn missing_ed25519_issuer_created() {
         None,
         None,
         Some(BECH32_ADDRESS_ED25519_1),
+        None,
         None,
     )]);
 
@@ -505,6 +546,7 @@ fn missing_ed25519_issuer_transition() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -513,6 +555,7 @@ fn missing_ed25519_issuer_transition() {
         None,
         None,
         Some(BECH32_ADDRESS_ED25519_1),
+        None,
         None,
     )]);
 
@@ -540,6 +583,7 @@ fn missing_alias_sender() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -547,6 +591,7 @@ fn missing_alias_sender() {
         BECH32_ADDRESS_ED25519_0,
         None,
         Some(BECH32_ADDRESS_ALIAS_1),
+        None,
         None,
         None,
     )]);
@@ -578,6 +623,7 @@ fn missing_alias_issuer_created() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -586,6 +632,7 @@ fn missing_alias_issuer_created() {
         None,
         None,
         Some(BECH32_ADDRESS_ALIAS_1),
+        None,
         None,
     )]);
 
@@ -616,6 +663,7 @@ fn missing_alias_issuer_transition() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -624,6 +672,7 @@ fn missing_alias_issuer_transition() {
         None,
         None,
         Some(BECH32_ADDRESS_ALIAS_1),
+        None,
         None,
     )]);
 
@@ -651,6 +700,7 @@ fn missing_nft_sender() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -658,6 +708,7 @@ fn missing_nft_sender() {
         BECH32_ADDRESS_ED25519_0,
         None,
         Some(BECH32_ADDRESS_NFT_1),
+        None,
         None,
         None,
     )]);
@@ -689,6 +740,7 @@ fn missing_nft_issuer_created() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -697,6 +749,7 @@ fn missing_nft_issuer_created() {
         None,
         None,
         Some(BECH32_ADDRESS_NFT_1),
+        None,
         None,
     )]);
 
@@ -727,6 +780,7 @@ fn missing_nft_issuer_transition() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
@@ -735,6 +789,7 @@ fn missing_nft_issuer_transition() {
         None,
         None,
         Some(BECH32_ADDRESS_NFT_1),
+        None,
         None,
     )]);
 
@@ -755,13 +810,23 @@ fn increase_nft_amount() {
     let nft_id_1 = NftId::from_str(NFT_ID_1).unwrap();
 
     let inputs = build_inputs(vec![
-        Nft(2_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
-        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None),
+        Nft(
+            2_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
     let outputs = build_outputs(vec![Nft(
         3_000_000,
         nft_id_1,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -787,13 +852,23 @@ fn decrease_nft_amount() {
     let nft_id_1 = NftId::from_str(NFT_ID_1).unwrap();
 
     let inputs = build_inputs(vec![
-        Nft(2_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
-        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None),
+        Nft(
+            2_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
     let outputs = build_outputs(vec![Nft(
         1_000_000,
         nft_id_1,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -831,12 +906,22 @@ fn prefer_basic_to_nft() {
     let nft_id_1 = NftId::from_str(NFT_ID_1).unwrap();
 
     let inputs = build_inputs(vec![
-        Nft(2_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
-        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None),
+        Nft(
+            2_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -864,12 +949,22 @@ fn take_amount_from_nft_to_fund_basic() {
     let nft_id_1 = NftId::from_str(NFT_ID_1).unwrap();
 
     let inputs = build_inputs(vec![
-        Nft(2_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
-        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None),
+        Nft(
+            2_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+        Basic(1_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
     ]);
     let outputs = build_outputs(vec![Basic(
         1_200_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -910,14 +1005,24 @@ fn nft_burn_should_validate_nft_sender() {
     let nft_id_1 = NftId::from_str(NFT_ID_1).unwrap();
 
     let inputs = build_inputs(vec![
-        Basic(2_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None),
-        Nft(1_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
+        Basic(2_000_000, BECH32_ADDRESS_ED25519_0, None, None, None, None, None, None),
+        Nft(
+            1_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
     ]);
     let outputs = build_outputs(vec![Basic(
         3_000_000,
         BECH32_ADDRESS_ED25519_0,
         None,
         Some(BECH32_ADDRESS_NFT_1),
+        None,
         None,
         None,
         None,
@@ -943,12 +1048,22 @@ fn nft_burn_should_validate_nft_address() {
     let nft_id_1 = NftId::from_str(NFT_ID_1).unwrap();
 
     let inputs = build_inputs(vec![
-        Basic(2_000_000, BECH32_ADDRESS_NFT_1, None, None, None, None, None),
-        Nft(1_000_000, nft_id_1, BECH32_ADDRESS_ED25519_0, None, None, None, None),
+        Basic(2_000_000, BECH32_ADDRESS_NFT_1, None, None, None, None, None, None),
+        Nft(
+            1_000_000,
+            nft_id_1,
+            BECH32_ADDRESS_ED25519_0,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
     ]);
     let outputs = build_outputs(vec![Basic(
         3_000_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
@@ -983,10 +1098,12 @@ fn transitioned_zero_nft_id_no_longer_is_zero() {
         None,
         None,
         None,
+        None,
     )]);
     let outputs = build_outputs(vec![Basic(
         1_000_000,
         BECH32_ADDRESS_ED25519_0,
+        None,
         None,
         None,
         None,
