+++ client/tests/signing/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod alias;
+mod basic;
